feat: organize icons into a nested folder structure
This commit reorganizes the icons within the `icons` folder by creating a new folder called `misc` and nesting relevant icons inside it. The new folder structure enhances the organization and maintainability of the icons, making it easier to locate and manage social-related icons in the future.

No changes were made to the actual icon files themselves, only their placement within the folder structure was adjusted.

The new folder structure is as follows:
- `assets/images/icons/misc/`: Contains miscellaneous icons.

This reorganization improves the overall structure of the project and promotes better organization of assets.
